refactor: metadata entity
