+++ server/src/types.ts
@@ -11,6 +11,8 @@ import {
   SyncEntityType,
   SystemMetadataKey,
   TranscodeTarget,
+  UserAvatarColor,
+  UserMetadataKey,
   VideoCodec,
 } from 'src/enum';
 
@@ -454,4 +456,55 @@ export interface SystemMetadata extends Record<SystemMetadataKey, Record<string,
   [SystemMetadataKey.SYSTEM_FLAGS]: DeepPartial<SystemFlags>;
   [SystemMetadataKey.VERSION_CHECK_STATE]: VersionCheckMetadata;
   [SystemMetadataKey.MEMORIES_STATE]: MemoriesState;
+}
+
+export type UserMetadataItem<T extends keyof UserMetadata = UserMetadataKey> = {
+  key: T;
+  value: UserMetadata[T];
+};
+
+export interface UserPreferences {
+  folders: {
+    enabled: boolean;
+    sidebarWeb: boolean;
+  };
+  memories: {
+    enabled: boolean;
+  };
+  people: {
+    enabled: boolean;
+    sidebarWeb: boolean;
+  };
+  ratings: {
+    enabled: boolean;
+  };
+  sharedLinks: {
+    enabled: boolean;
+    sidebarWeb: boolean;
+  };
+  tags: {
+    enabled: boolean;
+    sidebarWeb: boolean;
+  };
+  avatar: {
+    color: UserAvatarColor;
+  };
+  emailNotifications: {
+    enabled: boolean;
+    albumInvite: boolean;
+    albumUpdate: boolean;
+  };
+  download: {
+    archiveSize: number;
+    includeEmbeddedVideos: boolean;
+  };
+  purchase: {
+    showSupportBadge: boolean;
+    hideBuyButtonUntil: string;
+  };
+}
+
+export interface UserMetadata extends Record<UserMetadataKey, Record<string, any>> {
+  [UserMetadataKey.PREFERENCES]: DeepPartial<UserPreferences>;
+  [UserMetadataKey.LICENSE]: { licenseKey: string; activationKey: string; activatedAt: string };
 }
